Fixes #304 by setting random seeds (#314)
* Fixes #304 by setting random seeds

* Fix pylint and mypy
